+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.6.3] - 2022-10-25
 
 ### Fixed
 
@@ -230,3 +230,4 @@ _See [#25](https://github.com/attackgoat/screen-13/pull/25) for migration detail
 [0.6.0]: https://crates.io/crates/screen-13/0.6.0
 [0.6.1]: https://crates.io/crates/screen-13/0.6.1
 [0.6.2]: https://crates.io/crates/screen-13/0.6.2
+[0.6.3]: https://crates.io/crates/screen-13/0.6.3